requirements.txt
gspread
streamlit_authenticator
streamlit_cookies_controller
slack_sdk
anthropic
python-dotenv